+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `glsl-lang`: Add `serde` feature for AST serialization
 - `glsl-lang-cli`: Add JSON output format
 - `glsl-lang-cli`: Add input file argument
+- `glsl-lang-cli`: Add GLSL output format
 
 ### Changed
 - PartialEq for Node now compares the contents of the node by default